Support batched packages
Packages in a batch share the same version. This is designed to ease
tracking of subpackages (qemu-system-x86_64 and qemu-system-aarch64) and
group packages (all KDE packages come with the same version).

A new package property "follow" is added to indicate a batched package
and specify source package for synchronizing its version information.

Signed-off-by: Yao Zi <ziyao@disroot.org>